api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.36
asyncpg==0.30.0
pyjwt==2.8.0
python-multipart==0.0.6
pydantic-settings==2.1.0
httpx==0.25.2
greenlet>=3.1.0
